learning-python/fastapi-supabase/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
supabase
python-decouple
python-dotenv